requirements.txt
flask==3.0.3
flask-cors==4.0.1
google-generativeai==0.7.2
PyJWT==2.9.0
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.2
Werkzeug==3.0.4
requests==2.32.3
